+++ fabric/src/main/java/top/vmctcn/vmtranslationupdate/fabric/VMTranslationUpdateClientFabric.java
@@ -5,7 +5,7 @@ import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.screen.v1.ScreenEvents;
 import top.vmctcn.vmtranslationupdate.ModEvents;
 import top.vmctcn.vmtranslationupdate.VMTranslationUpdate;
-import top.vmctcn.vmtranslationupdate.screen.ScreenAfterInitEvent;
+import top.vmctcn.vmtranslationupdate.util.ScreenUtil;
 import top.vmctcn.vmtranslationupdate.util.ModConfigUtil;
 
 public class VMTranslationUpdateClientFabric implements ClientModInitializer {
@@ -20,7 +20,7 @@ public class VMTranslationUpdateClientFabric implements ClientModInitializer {
         });
 
         ScreenEvents.AFTER_INIT.register((client, screen, scaledWidth, scaledHeight) -> {
-            ScreenAfterInitEvent.screenAfterInitEvent(screen);
+            ScreenUtil.screenAfterInitEvent(screen);
         });
     }
 }